Move clients to separate namespace
